修改

+++ Assets/Scripts/General/Attack.cs
@@ -4,12 +4,12 @@ using UnityEngine;
 
 public class Attack : MonoBehaviour
 {
-    public int damage;
+    public float damage;
     public float attackRange;
     public float attackRate;
     private void OnTriggerStay2D(Collider2D other)
     {
         other.GetComponent<Character>()?.TakeDamage(this);
     }
-
+    
 }
